Week4/v4/server/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyJWT
